editing a cabin

+++ src/services/apiCabins.js
@@ -11,35 +11,46 @@ export async function getCabins() {
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replace("/", "");
+export async function createEditCabin(newCabin, id) {
+  console.log(newCabin);
 
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
 
-  // https://ybsucqavvwiwqmwxtnfw.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
+  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
+    "/",
+    ""
+  );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
 
-  // 1. create cabin
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
-    console.error("error");
-    throw new Error("Cabins could not be created");
+    console.error(error);
+    throw new Error("Cabin could not be created");
   }
 
-  // 2. upload image
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
 
-  // 3. delete the cabin if there was an error uploading an image
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
-
     console.error(storageError);
-
     throw new Error(
       "Cabin image could not be uploaded and the cabin was not created"
     );
